+++ src/main/java/camp/nextstep/edu/racingcar/Main.java
@@ -1,5 +1,11 @@
 package camp.nextstep.edu.racingcar;
 
+import camp.nextstep.edu.racingcar.view.RacingGameConsole;
+import camp.nextstep.edu.racingcar.view.RacingInputView;
+import camp.nextstep.edu.racingcar.view.RacingInputViewImpl;
+import camp.nextstep.edu.racingcar.view.RacingOutputView;
+import camp.nextstep.edu.racingcar.view.RacingOutputViewImpl;
+
 public class Main {
 
     public static void main(String[] args) {
