+++ src/main/java/camp/nextstep/edu/racingcar/domain/Cars.java
@@ -1,4 +1,8 @@
-package camp.nextstep.edu.racingcar;
+package camp.nextstep.edu.racingcar.domain;
+
+import camp.nextstep.edu.racingcar.domain.result.DriveResult;
+import camp.nextstep.edu.racingcar.domain.strategy.DriveStrategy;
+import camp.nextstep.edu.racingcar.domain.result.RoundResult;
 
 public class Cars {
