+++ src/main/java/camp/nextstep/edu/racingcar/domain/strategy/DriveStrategy.java
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.racingcar;
+package camp.nextstep.edu.racingcar.domain.strategy;
 
 public interface DriveStrategy {
