+++ src/test/java/camp/nextstep/edu/racingcar/CarTest.java
@@ -2,6 +2,9 @@ package camp.nextstep.edu.racingcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import camp.nextstep.edu.racingcar.domain.Car;
+import camp.nextstep.edu.racingcar.domain.result.DriveResult;
+import camp.nextstep.edu.racingcar.domain.strategy.DriveStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
