+++ src/test/java/camp/nextstep/edu/racingcar/RacingGameTest.java
@@ -2,6 +2,9 @@ package camp.nextstep.edu.racingcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import camp.nextstep.edu.racingcar.domain.RacingGame;
+import camp.nextstep.edu.racingcar.domain.result.RacingGameResult;
+import camp.nextstep.edu.racingcar.domain.strategy.DriveStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
